Introduce ElemRef for QualifiedNameableElem elements
We introduce ElemRef that acts as a reference indirection to elements that may be
lazy loaded.

We introduce the QualifiedNameableElem abstract class that derives from Elem
and implements jx.l.m.e.QualifiedNameable.

PackageElem and TypeElem now extends QualifiedNameableElem.

We only use the indirection in the DeclaredTpe to TypeElem relationship. Indeed,
the enclosed element relationship must be set up at construction time thus cannot
use lazy-loading. This may not be a problem anyway because when the semantic
analysis needs an inner-class type to resolve some names, it may just after need
its outer-class as the type elem scope dictates.

Thus, for now, loading an inner-class from class file requires to first load its
outer-class and the inner-classes of this class (and recursively up to top-level).
We may decide to change that later for performance reasons if really necessary.